+++ include/commands/CommandSound.h
@@ -10,7 +10,13 @@
 
 #include "commands/Command.h"
 
-DECLARE_SIMPLE_CMD(CommandSound);
+class CommandSound : public Command {
+public:
+    virtual std::string name();
+    virtual std::string brief();
+    virtual void printHelp();
+    virtual int execute(std::istream& args);
+};
 
 #endif